Refactor: Update SlackClient import in slack_notify.py
